Change: Disable jumpping out of delimiters by right delimiter.
This is a very primitive, inconsiderate modification,  to
address the following need:

When the cursor is inside an empty pair of quotes/delimiters
or just before the right quote/delimiter, typing THAT kind
of delimiter should never trigger a jumpout. Rather, it should
insert another pair of quotes/delimiters of request (for empty
quotes/delimiters) or insert one quote/delimiter of request
(for smart-quote, etc.).

In this way, we unified the jumpout operation to be performed
solely by `DelimitMateS-Tab`. And let quote/delimiter chars
always insert characters and never jump unexpectedly.